+++ ncbi_submit/xml_format.py
@@ -4,6 +4,8 @@
 import datetime
 from abc import ABC, abstractmethod
 from textwrap import indent,dedent
+from typing import Literal
+import pandas as pd
 from ncbi_submit.helpers import series2dict,get_bioproject_spuid
 # from ncbi_submit.ncbi import NCBI # only use for testing (with typing hints) - otherwise causes circular import
 
@@ -11,7 +13,12 @@ class Action(ABC):
     """Converts details about an individual sample to an XML action"""
 
     def __init__(self,ncbi,indent_by=2) -> None:
-        """An object that can generate the text for an xml submission <Action> block"""
+        """An object that can generate the text for an xml submission <Action> block
+        
+        Attributes:
+            `ncbi` (NCBI): An object storing information about the submission data, goals, and useful methods
+            `indent_by` (int): Number of spaces by which to indent these lines. Defaults to 2.
+        """
 
         ncbi:NCBI
         self.ncbi = ncbi
@@ -47,8 +54,16 @@ class GenBank_Action(Action):
 
 class SRA_BioSample_Action(Action):
 
-    def __init__(self,action_type,ncbi,biosample_link,indent_by=2,attributes={}) -> None:
-        """Instantiates SRA_BioSample_Action"""
+    def __init__(self,action_type:Literal["sra","biosample"],ncbi,biosample_link,indent_by=2,attributes={}) -> None:
+        """Instantiates SRA_BioSample_Action
+        
+        Attributes
+            `action_type` (str): Database that this action is for.
+            `ncbi` (NCBI): An object storing information about the submission data, goals, and useful methods
+            `biosample_link` (str): xml tag line linking to BioSample (spuid or accession)
+            `indent_by` (int): Number of spaces by which to indent these lines. Defaults to 2.
+            `attributes` (dict): Attributes to go in the xml attributes sections
+        """
 
         ncbi:NCBI
         self.action_type = action_type
@@ -84,7 +99,12 @@ class SRA_BioSample_Action(Action):
             yield f'<Attribute {name_of_attr}="{attr_name}">{attr_value}</Attribute>'
 
     def get_bioproject_link(self,accession=None):
-        """Returns link to bioproject"""
+        """Returns link to bioproject
+        
+        Args:
+            `accession` (str,optional): The bioproject accession to link the submission to.
+        """
+
         if self.ncbi.bioproject_presets["create_new"]:
             # link to BioProject SPUID
             return get_bioproject_spuid(self.ncbi)
@@ -96,6 +116,7 @@ class SRA_BioSample_Action(Action):
 
     def reference_bioproject(self): # this could be replaced by a function that only links biosample if not creating it as a seperate Action step
         """Returns xml text for referencing BioProject accession"""
+
         return dedent(f"""\
         <AttributeRefId name="BioProject">
           <RefId>
@@ -106,16 +127,30 @@ class SRA_BioSample_Action(Action):
 class SRA_Action(SRA_BioSample_Action):
     """A class to produce SRA or BioSample XML actions"""
 
-    def __init__(self,ncbi,biosample_link,sra_link,indent_by=2,attributes={},accessions_dict=None) -> None:
-        """Instantiates SRA_Action"""
+    def __init__(self,ncbi,biosample_link,sra_link,indent_by=2,attributes={},accession_dict=None) -> None:
+        """Instantiates SRA_Action
+        
+        Attributes:
+            `ncbi` (NCBI): An object storing information about the submission data, goals, and useful methods
+            `biosample_link` (str): xml tag line linking to BioSample (spuid or accession)
+            `sra_link` (str): xml tag line containing SRA spuid
+            `indent_by` (int): Number of spaces by which to indent these lines. Defaults to 2.
+            `attributes` (dict): Attributes to go in the xml attributes sections
+            `accession_dict` (dict): maps sample_name -> biosample accession. If sample already has accession, xml will link to accession otherwise, uses spuid
+        """
 
         super().__init__("sra",ncbi=ncbi,biosample_link=biosample_link,indent_by=indent_by,attributes=attributes)
         self.sra_link = sra_link
-        self.accessions_dict = accessions_dict
+        self.accession_dict = accession_dict
         self.ncbi:NCBI
 
     def link_bioSample(self,biosample_spuid):
-        "returns xml text referencing corresponding BioSample (for SRA submission)"
+        """Returns SRA xml text referencing corresponding BioSample
+
+        Args:
+          `biosample_spuid` (str): spuid or biosample accession tag line
+        """
+
         return dedent(f"""\
         <AttributeRefId name="BioSample">
           <RefId>
@@ -164,7 +199,14 @@ class BioSample_Action(SRA_BioSample_Action):
     """A class to produce SRA XML actions"""
 
     def __init__(self,ncbi,biosample_link,indent_by=2,attributes={}) -> None:
-        """Instantiates BioSample_Action"""
+        """Instantiates BioSample_Action
+        
+        Attributes
+            `ncbi` (NCBI): An object storing information about the submission data, goals, and useful methods
+            `biosample_link` (str): xml tag line linking to BioSample (spuid or accession)
+            `indent_by` (int): Number of spaces by which to indent these lines. Defaults to 2.
+            `attributes` (dict): Attributes to go in the xml attributes sections
+        """
 
         super().__init__("biosample",ncbi=ncbi,biosample_link=biosample_link,indent_by=indent_by,attributes=attributes)
 
@@ -227,18 +269,19 @@ class Submission(ABC):
     """Converts TSV to XML submission"""
 
     # @abstractmethod
-    def __init__(self,ncbi,data_type) -> None:
-    # def __init__(self,plate,data_type:str,biosample_presets:dict,affiliation,email,contact,centerAbbr,test_dir=False) -> None:
-        """Instantiate a Submission object
+    def __init__(self,ncbi,data_type:Literal["BS+SRA","BS","SRA","GB"],**kwargs) -> None:
+        """Abstract class for creating submission xml files
         
         Attributes:
-            ncbi (NCBI): An object storing information about the submission data and goals
-            data_type (str): name of submission type ("BS+SRA","GB","BS","SRA")
+            `ncbi` (NCBI): An object storing information about the submission data, goals, and useful methods
+            `data_type` (str): name of submission type
         """
-        
+
         ncbi:NCBI
         self.ncbi = ncbi
         self.data_type = data_type
+        # for key,val in kwargs.items():
+        #     setattr(self,key,val)
 
     @abstractmethod
     def xml_actions(self):
@@ -253,9 +296,7 @@ class Submission(ABC):
         else: return ""
 
     def xml_description(self):
-        """Returns lab-specific header for any NCBI xml file
-        
-        """
+        """Returns lab-specific header for any NCBI xml file"""
 
         ## TODO: allow date specifications in `config_file`?
         # next_week = datetime.datetime.now() + datetime.timedelta(days=7)
@@ -290,7 +331,7 @@ class Submission(ABC):
         yield "</Submission>"
 
     def write_xml(self,filename):
-        """Combines all XML pieces and writes to '`outfile`/`filename`'"""
+        """Combines all XML pieces and writes to '`filename`'"""
 
         with filename.open("w") as out:
             for line in self.generate_xml_lines():
@@ -312,36 +353,82 @@ class GenBank_Submission(Submission):
 class SRA_BioSample_Submission(Submission):
     """Converts SRA and BioSample TSVs to submission XML"""
 
-    def __init__(self,ncbi,add_biosample=True,add_sra=True,data_type="BS+SRA") -> None:
+    def __init__(self,ncbi,add_biosample=True,add_sra=True,update_reads=False,spuid_endings=None,report_files=[],download_reports=True) -> None:
         """Instantiate a Submission object
         
         Attributes:
-            ncbi (NCBI): An object storing information about the submission data and goals
-            data_type (str): name of submission type ("BS+SRA","GB") # TODO: remove?
-            db (str): database to be submitted to ("bs_sra","gb")
-            biosample_df (DataFrame): df containing final data for biosample_df submission
-            sra_df (DataFrame): df containing final data for sra_df submission
+            `ncbi` (NCBI): An object storing information about the submission data, goals, and useful methods
+            `add_biosample` (bool): whether to add biosample xml actions to the submission
+            `add_sra` (bool): whether to add sra xml actions to the submission
+            `update_reads` (bool): whether this submission includes any samples for which the reads are being updated (were previously submitted)
+            `spuid_endings` (str | dict): SPUID suffixes to use for specified samples.
+              (as str): 'suffix1:samp1,samp2;suffix2:samp3'. Strings will be converted to the dict format.
+              (as dict): {samp1:suffix1,samp2:suffix1,samp3:suffix2}
+            `biosample_df` (DataFrame): df containing final data for biosample_df submission
+            `sra_df` (DataFrame): df containing final data for sra_df submission
+            `report_files` (list, optional): List of report files from which to seek accessions, if provided. Defaults to None.
+            `download_reports` (bool, optional): If True and `update_reads` is True, reports will be downloaded. Defaults to False.
+        Auto-set attributes:
+            `data_type` (str): name of submission type ("BS+SRA") # TODO: remove?
+            `db` (str): database for which to collect report info ("bs_sra")
         """
 
-        super().__init__(ncbi,data_type=data_type)
         ncbi:NCBI
-        # self.ncbi = ncbi
-        # self.data_type = data_type # TODO: remove?
+        if update_reads:
+            if not ncbi.ftp:
+                ncbi.ncbiConnect()
+            ncbi._get_all_accessions("bs_sra",report_files=report_files,download_reports=download_reports)
+        super().__init__(ncbi,data_type="BS+SRA")
         self.add_biosample = add_biosample
         self.add_sra = add_sra
         self.db = "bs_sra"
+        self.update_reads = update_reads
+        self.spuid_endings = self.parseSpuidEndings(spuid_endings)
+
+    def parseSpuidEndings(self,spuid_endings=None):
+        """Converts user-provided `spuid_endings` to a dictionary
+
+        Args:
+            `spuid_endings` (str | dict): SPUID suffixes to use for specified samples.
+              (as str): 'suffix1:samp1,samp2;suffix2:samp3'. Strings will be converted to the dict format.
+              (as dict): {samp1:suffix1,samp2:suffix1,samp3:suffix2}
+        """
+
+        if not spuid_endings:
+            return {}
+        if type(spuid_endings) is dict:
+            return spuid_endings
+        elif type(spuid_endings) is str:
+            spuid_endings = spuid_endings.strip("'\"")
+            endings_dict = {}
+            suffix_groups = spuid_endings.split(";")
+            for group in suffix_groups:
+                suffix = group.split(":")[0]
+                samples = group.split(":")[1].split(",")
+                for sample in samples:
+                    endings_dict[sample] = suffix
+            return endings_dict
+        else: raise TypeError("`spuid_endings` must be a string or a dictionary")
 
     def getSpuidOrLink(self,row,sra_only,accession_dict):
-        """Returns SPUID or link for SRA and BioSample submissions"""
+        """Returns SPUID or link for SRA and BioSample submissions
+        
+        Args:
+            `row` (pd.Series|dict): sample-specific details
+            `sra_only` (bool): whether samples are only being submitted to SRA this time
+            `accession_dict` (dict): maps sample_name -> biosample accession. If sample already has accession, xml will link to accession otherwise, uses spuid
+        """
 
         num = f"_{self.ncbi.attemtpt_num}" if self.ncbi.vary_spuid else ""
-        sra_link = f"""<SPUID spuid_namespace="{self.ncbi.centerAbbr}">{row["sample_name"]}_SRA{num}</SPUID>"""
-        if sra_only:
+        # print(row["sample_name"])
+        suffix = self.spuid_endings.get(row["sample_name"],"")
+        sra_link = f"""<SPUID spuid_namespace="{self.ncbi.centerAbbr}">{row["sample_name"]}_SRA{num}{suffix}</SPUID>"""
+        if sra_only or (self.update_reads and row["sample_name"] in accession_dict.keys()):
             # assuming BioSamples were previously submitted, link to their accessions, not their name
             biosample_link = f"""<PrimaryId db="BioSample">{accession_dict[row["sample_name"]]}</PrimaryId>"""
             # xml_text = self.sra_action(row[sra_cols],sra_link,biosample_link,bioproject_accession) # temp
         else:
-            biosample_link = f"""<SPUID spuid_namespace="{self.ncbi.centerAbbr}">{row["sample_name"]}_BioSample{num}</SPUID>"""
+            biosample_link = f"""<SPUID spuid_namespace="{self.ncbi.centerAbbr}">{row["sample_name"]}_BioSample{num}{suffix}</SPUID>"""
             # xml_text = self.biosample_action(row[biosample_cols],biosample_link,plate,test_dir) + "\n" + self.sra_action(row[sra_cols],sra_link,biosample_link,bioproject_accession)
         return sra_link,biosample_link
 
@@ -359,14 +446,18 @@ class SRA_BioSample_Submission(Submission):
         # required if only submitting SRA (and BioSample has already been submitted)
         # TODO: test with `sra_only`
         if self.add_biosample and not self.add_sra:
-            accessions_dict = self.ncbi.get_accessions(as_dict=True)
-        else: accessions_dict = None
+            accession_dict = self.ncbi.get_accessions(as_dict=True)
+        elif self.update_reads:
+            acc_df = self.ncbi.sra_bs_accessions_df[["sample_name","BioSample"]]
+            acc_df = acc_df[acc_df["sample_name"].isin(self.ncbi.sra["df"]["sample_name"])]
+            accession_dict = dict(zip(acc_df["sample_name"],acc_df["BioSample"]))
+            if len(accession_dict)==0: raise AttributeError(f"No previously-submitted BioSample accessions found but the flag --update_reads was used. \nTo specify reports with those accessions, use the --report_files flag. \nAccessions can also be written directly in a csv and pointed to via the variable `extra_accessions` (currently {self.ncbi.extra_accessions}) in your config file (currently {self.ncbi.config_file})")
+        else: accession_dict = {}
 
         # alternate adding BioSample and SRA actions for each sample
         for i,row in self.ncbi.merged["df"].iterrows():
-
-            sra_link,biosample_link = self.getSpuidOrLink(row, not self.add_biosample, accessions_dict)
-            if self.add_biosample:
+            sra_link,biosample_link = self.getSpuidOrLink(row=row, sra_only=not self.add_biosample, accession_dict=accession_dict)
+            if self.add_biosample and not (self.update_reads and row["sample_name"] in accession_dict.keys()):
                 for line in BioSample_Action(
                     ncbi=self.ncbi,
                     attributes=series2dict(row,biosample_cols),
@@ -379,13 +470,15 @@ class SRA_BioSample_Submission(Submission):
                     attributes=series2dict(row,sra_cols),
                     biosample_link=biosample_link,
                     sra_link=sra_link,
-                    accessions_dict=accessions_dict,
+                    accession_dict=accession_dict,
                     ).generate_action_lines():
                     yield line
 
     def add_BioProject_xml(self):
-        "returns xml action to create a new BioProject"
-        # see https://www.ncbi.nlm.nih.gov/viewvc/v1/trunk/submit/public-docs/bioproject/bioproject.xsd?view=co
+        """Returns xml action to create a new BioProject
+        For more, see [NCBI BioProject Info](https://www.ncbi.nlm.nih.gov/viewvc/v1/trunk/submit/public-docs/bioproject/bioproject.xsd?view=co)
+        
+        """
         bioproject_spuid = get_bioproject_spuid(self.ncbi)
         return indent(dedent(f"""\
         <Action>